# Check for positive definiteness only after computing the LLT factors