--- a/stan/math/prim/err/check_pos_definite.hpp
+++ b/stan/math/prim/err/check_pos_definite.hpp
@@ -15,14 +15,19 @@ namespace math {
 
 /**
  * Check if the specified square, symmetric matrix is positive definite.
- * @tparam T_y Type of scalar of the matrix
- * @param function Function name (for error messages)
- * @param name Variable name (for error messages)
- * @param y Matrix to test
- * @throw <code>std::invalid_argument</code> if the matrix is not square
+ *
+ * This computes an LDLT decomposition to establish positive definiteness,
+ * so it can be expensive. If an LDLT or LLT decomposition is available,
+ * that should be tested instead.
+ *
+ * @tparam T_y type of elements in the matrix
+ * @param function function name (for error messages)
+ * @param name variable name (for error messages)
+ * @param y matrix to test
+ * @throw std::invalid_argument if the matrix is not square
  * or if the matrix has 0 size.
- * @throw <code>std::domain_error</code> if the matrix is not symmetric,
- * if it is not positive definite, or if any element is <code>NaN</code>
+ * @throw std::domain_error if the matrix is not symmetric,
+ * if it is not positive definite, or if any element is NaN
  */
 template <typename T_y>
 inline void check_pos_definite(const char* function, const char* name,
@@ -43,14 +48,14 @@ inline void check_pos_definite(const char* function, const char* name,
 }
 
 /**
- * Check if the specified LDLT transform of a matrix is positive definite.
- * @tparam Derived Derived type of the Eigen::LDLT transform.
- * @param function Function name (for error messages)
- * @param name Variable name (for error messages)
+ * Check if the specified LDLT decomposition of a matrix is positive definite.
+ *
+ * @tparam Derived type of the Eigen::LDLT decomposition
+ * @param function function name (for error messages)
+ * @param name variable name (for error messages)
  * @param cholesky Eigen::LDLT to test, whose progenitor
  * must not have any NaN elements
- * @throw <code>std::domain_error</code> if the matrix is not
- * positive definite
+ * @throw std::domain_error if the matrix is not positive definite
  */
 template <typename Derived>
 inline void check_pos_definite(const char* function, const char* name,
@@ -62,15 +67,15 @@ inline void check_pos_definite(const char* function, const char* name,
 }
 
 /**
- * Check if the specified LLT decomposition transform resulted in
- * <code>Eigen::Success</code>
- * @tparam Derived Derived type of the Eigen::LLT transform.
- * @param function Function name (for error messages)
- * @param name Variable name (for error messages)
+ * Check if the specified LLT decomposition was successful.
+ *
+ * @tparam Derived type of the Eigen::LLT decomposition
+ * @param function function name (for error messages)
+ * @param name variable name (for error messages)
  * @param cholesky Eigen::LLT to test, whose progenitor
  * must not have any NaN elements
- * @throw <code>std::domain_error</code> if the diagonal of the
- * L matrix is not positive
+ * @throw std::domain_error if the decomposition failed or the
+ * diagonal of the L matrix is not positive
  */
 template <typename Derived>
 inline void check_pos_definite(const char* function, const char* name,