--- a/stan/math/prim/fun/mdivide_left_spd.hpp
+++ b/stan/math/prim/fun/mdivide_left_spd.hpp
@@ -30,7 +30,7 @@ namespace math {
 template <typename T1, typename T2, int R1, int C1, int R2, int C2>
 inline Eigen::Matrix<return_type_t<T1, T2>, R1, C2> mdivide_left_spd(
     const Eigen::Matrix<T1, R1, C1> &A, const Eigen::Matrix<T2, R2, C2> &b) {
-  static const char* function = "mdivide_left_spd";
+  static const char *function = "mdivide_left_spd";
   check_multiplicable(function, "A", A, "b", b);
   check_positive(function, "rows", A.rows());
   check_symmetric(function, "A", A);