--- a/stan/math/prim/fun/mdivide_right_spd.hpp
+++ b/stan/math/prim/fun/mdivide_right_spd.hpp
@@ -31,7 +31,7 @@ namespace math {
 template <typename T1, typename T2, int R1, int C1, int R2, int C2>
 inline Eigen::Matrix<return_type_t<T1, T2>, R1, C2> mdivide_right_spd(
     const Eigen::Matrix<T1, R1, C1> &b, const Eigen::Matrix<T2, R2, C2> &A) {
-  static const char* function = "mdivide_right_spd";
+  static const char *function = "mdivide_right_spd";
   check_multiplicable(function, "b", b, "A", A);
   check_positive(function, "rows", A.rows());
   check_symmetric(function, "A", A);