--- a/stan/math/rev/fun/mdivide_left_spd.hpp
+++ b/stan/math/rev/fun/mdivide_left_spd.hpp
@@ -51,6 +51,7 @@ class mdivide_left_spd_vv_vari : public vari {
     Eigen::Map<matrix_vi>(variRefB_, M_, N_) = B.vi();
     alloc_->C_ = B.val();
     alloc_->llt_ = A.val().llt();
+    check_pos_definite("mdivide_left_spd", "A", alloc_->llt_);
     alloc_->llt_.solveInPlace(alloc_->C_);
 
     Eigen::Map<matrix_vi>(variRefC_, M_, N_)
@@ -90,6 +91,7 @@ class mdivide_left_spd_dv_vari : public vari {
     alloc_->C_ = B.val();
     Eigen::Map<matrix_vi>(variRefB_, M_, N_) = B.vi();
     alloc_->llt_ = A.llt();
+    check_pos_definite("mdivide_left_spd", "A", alloc_->llt_);
     alloc_->llt_.solveInPlace(alloc_->C_);
 
     Eigen::Map<matrix_vi>(variRefC_, M_, N_)
@@ -126,6 +128,7 @@ class mdivide_left_spd_vd_vari : public vari {
         alloc_(new mdivide_left_spd_alloc<R1, C1, R2, C2>()) {
     Eigen::Map<matrix_vi>(variRefA_, M_, M_) = A.vi();
     alloc_->llt_ = A.val().llt();
+    check_pos_definite("mdivide_left_spd", "A", alloc_->llt_);
     alloc_->C_ = alloc_->llt_.solve(B);
 
     Eigen::Map<matrix_vi>(variRefC_, M_, N_)
@@ -144,10 +147,11 @@ template <int R1, int C1, int R2, int C2>
 inline Eigen::Matrix<var, R1, C2> mdivide_left_spd(
     const Eigen::Matrix<var, R1, C1> &A, const Eigen::Matrix<var, R2, C2> &b) {
   Eigen::Matrix<var, R1, C2> res(b.rows(), b.cols());
-
-  check_square("mdivide_left_spd", "A", A);
-  check_multiplicable("mdivide_left_spd", "A", A, "b", b);
-  check_pos_definite("mdivide_left_spd", "A", A);
+  static const char* function = "mdivide_left_spd";
+  check_multiplicable(function, "A", A, "b", b);
+  check_positive(function, "rows", A.rows());
+  check_symmetric(function, "A", A);
+  check_not_nan(function, "A", A);
 
   // NOTE: this is not a memory leak, this vari is used in the
   // expression graph to evaluate the adjoint, but is not needed
@@ -164,11 +168,11 @@ template <int R1, int C1, int R2, int C2>
 inline Eigen::Matrix<var, R1, C2> mdivide_left_spd(
     const Eigen::Matrix<var, R1, C1> &A,
     const Eigen::Matrix<double, R2, C2> &b) {
-  Eigen::Matrix<var, R1, C2> res(b.rows(), b.cols());
-
-  check_square("mdivide_left_spd", "A", A);
-  check_multiplicable("mdivide_left_spd", "A", A, "b", b);
-  check_pos_definite("mdivide_left_spd", "A", A);
+  static const char* function = "mdivide_left_spd";
+  check_multiplicable(function, "A", A, "b", b);
+  check_positive(function, "rows", A.rows());
+  check_symmetric(function, "A", A);
+  check_not_nan(function, "A", A);
 
   // NOTE: this is not a memory leak, this vari is used in the
   // expression graph to evaluate the adjoint, but is not needed
@@ -177,6 +181,7 @@ inline Eigen::Matrix<var, R1, C2> mdivide_left_spd(
   internal::mdivide_left_spd_vd_vari<R1, C1, R2, C2> *baseVari
       = new internal::mdivide_left_spd_vd_vari<R1, C1, R2, C2>(A, b);
 
+  Eigen::Matrix<var, R1, C2> res(b.rows(), b.cols());
   res.vi() = Eigen::Map<matrix_vi>(&baseVari->variRefC_[0], b.rows(), b.cols());
   return res;
 }
@@ -185,11 +190,11 @@ template <int R1, int C1, int R2, int C2>
 inline Eigen::Matrix<var, R1, C2> mdivide_left_spd(
     const Eigen::Matrix<double, R1, C1> &A,
     const Eigen::Matrix<var, R2, C2> &b) {
-  Eigen::Matrix<var, R1, C2> res(b.rows(), b.cols());
-
-  check_square("mdivide_left_spd", "A", A);
-  check_multiplicable("mdivide_left_spd", "A", A, "b", b);
-  check_pos_definite("mdivide_left_spd", "A", A);
+  static const char* function = "mdivide_left_spd";
+  check_multiplicable(function, "A", A, "b", b);
+  check_positive(function, "rows", A.rows());
+  check_symmetric(function, "A", A);
+  check_not_nan(function, "A", A);
 
   // NOTE: this is not a memory leak, this vari is used in the
   // expression graph to evaluate the adjoint, but is not needed
@@ -198,6 +203,7 @@ inline Eigen::Matrix<var, R1, C2> mdivide_left_spd(
   internal::mdivide_left_spd_dv_vari<R1, C1, R2, C2> *baseVari
       = new internal::mdivide_left_spd_dv_vari<R1, C1, R2, C2>(A, b);
 
+  Eigen::Matrix<var, R1, C2> res(b.rows(), b.cols());
   res.vi() = Eigen::Map<matrix_vi>(&baseVari->variRefC_[0], b.rows(), b.cols());
 
   return res;