## Summary
This uses the `check_pos_definite` overload to check an LLT decompositions after it's computed, rather than using the generic `check_pos_definite` on the original matrix, as that would compute and then discard an LDLT decomposition of the matrix. Fixes #1693.

Since this other overload doesn't first check for NaNs or symmetry, I added those where necessary. Note that it's not necessary to check explicitly for a square matrix if we also check for symmetry, because the latter implies the former, that's why I removed some `check_square` in a couple of places.

I had a very brief look at the performance impact with this code:
```
TEST(MathMatrixPrim, mdivide_spd) {
  const int size = 100; // 200
  stan::math::vector_d d = Eigen::VectorXd::Ones(size) * 20;
  stan::math::matrix_d diag = d.asDiagonal();
  stan::math::matrix_d R = Eigen::MatrixXd::Random(size, size);
  stan::math::matrix_d A = 0.5 * (R + R.transpose()) + diag;

  for (int i = 0; i < 100; i++) {
    stan::math::matrix_d m1 = stan::math::mdivide_left_spd(A, A);
    stan::math::matrix_d m2 = stan::math::mdivide_right_spd(A, A);
  }
}
```
size 100: from 8706ms to 7011ms
size 200: from 49912ms to 40712ms

## Tests

No new tests.

## Side Effects

None.

## Checklist

- [X] Math issue #1693

- [X] Copyright holder: Marco Colombo

    The copyright holder is typically you or your assignee, such as a university or company. By submitting this pull request, the copyright holder is agreeing to the license the submitted work under the following licenses:
      - Code: BSD 3-clause (https://opensource.org/licenses/BSD-3-Clause)
      - Documentation: CC-BY 4.0 (https://creativecommons.org/licenses/by/4.0/)

- [X] the basic tests are passing

    - unit tests pass (to run, use: `./runTests.py test/unit`)
    - header checks pass, (`make test-headers`)
    - dependencies checks pass, (`make test-math-dependencies`)
    - docs build, (`make doxygen`)
    - code passes the built in [C++ standards](https://github.com/stan-dev/stan/wiki/Code-Quality) checks (`make cpplint`)

- [X] the code is written in idiomatic C++ and changes are documented in the doxygen

- [X] the new changes are tested